requirements.txt
# ═══════════════════════════════════════════════════════════
# Зависимости для Streamer Posts Bot
# ═══════════════════════════════════════════════════════════

# ────────────────────────────────────────────────────────────
# Telegram Bot (aiogram 3.x)
# ────────────────────────────────────────────────────────────
aiogram>=3.0.0

# ────────────────────────────────────────────────────────────
# Telegram Client (Telethon)
# ────────────────────────────────────────────────────────────
# Для работы с каналами (сканирование, копирование медиа)
telethon>=1.34.0

# ────────────────────────────────────────────────────────────
# AI APIs
# ────────────────────────────────────────────────────────────
# OpenAI SDK (используется для OpenRouter тоже)
openai>=1.0.0

# HTTP клиент для OpenRouter API
aiohttp>=3.9.0

# ────────────────────────────────────────────────────────────
# Конфигурация
# ────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ────────────────────────────────────────────────────────────
# Утилиты
# ────────────────────────────────────────────────────────────
# Асинхронность
asyncio-mqtt>=0.16.0

# Логирование
structlog>=23.0.0

# Системные утилиты
psutil>=5.9.0

# ═══════════════════════════════════════════════════════════
# УСТАНОВКА
# ═══════════════════════════════════════════════════════════
#
# pip install -r requirements.txt
#
# ═══════════════════════════════════════════════════════════
